Struct http_muncher::Parser

pub struct Parser { /* fields omitted */ }

The main parser interface.

Example

use http_muncher::{Parser, ParserHandler};
use std::str;

struct MyHandler;
impl ParserHandler for MyHandler {
   fn on_header_field(&mut self, parser: &mut Parser, header: &[u8]) -> bool {
       println!("{}: ", str::from_utf8(header).unwrap());
       true
   }
   fn on_header_value(&mut self, parser: &mut Parser, value: &[u8]) -> bool {
       println!("\t {:?}", str::from_utf8(value).unwrap());
       true
   }
}

let http_request = b"GET / HTTP/1.0\r\n\
                     Content-Length: 0\r\n\r\n";

let mut parser = Parser::request();
parser.parse(&mut MyHandler {}, http_request);

Methods

impl Parser

fn response() -> Parser

Creates a new parser instance for an HTTP response.

fn request() -> Parser

Creates a new parser instance for an HTTP request.

fn request_and_response() -> Parser

Creates a new parser instance to handle both HTTP requests and responses.

fn parse<H: ParserHandler>(&mut self, handler: &mut H, data: &[u8]) -> usize

Parses the provided data and returns a number of bytes read.

fn http_version(&self) -> (u16, u16)

Returns an HTTP request or response version.

fn status_code(&self) -> u16

Returns an HTTP response status code (think 404).

fn http_method(&self) -> &'static str

Returns an HTTP method static string (GET, POST, and so on).

fn has_error(&self) -> bool

Checks if the last parse call was finished successfully. Returns true if it wasn't.

fn error(&self) -> &'static str

In case of a parsing error returns its mnemonic name.

fn error_description(&self) -> &'static str

In case of a parsing error returns its description.

fn is_upgrade(&self) -> bool

Checks if an upgrade protocol (e.g. WebSocket) was requested.

fn is_final_chunk(&self) -> bool

Checks if it was the final body chunk.

fn should_keep_alive(&self) -> bool

If should_keep_alive() in the on_headers_complete or on_message_complete callback returns 0, then this should be the last message on the connection. If you are the server, respond with the "Connection: close" header. If you are the client, close the connection.

fn pause(&mut self)

fn unpause(&mut self)