Trait http_muncher::ParserHandler

pub trait ParserHandler {
    fn on_url(&self, &String) -> Option<u16> { ... }
    fn on_status(&self, &String) -> Option<u16> { ... }
    fn on_header_field(&self, &String) -> Option<u16> { ... }
    fn on_header_value(&self, &String) -> Option<u16> { ... }
    fn on_body(&self, &String) -> Option<u16> { ... }
    fn on_headers_complete(&self) -> Option<u16> { ... }
    fn on_message_begin(&self) -> Option<u16> { ... }
    fn on_message_complete(&self) -> Option<u16> { ... }
    fn on_chunk_header(&self) -> Option<u16> { ... }
    fn on_chunk_complete(&self) -> Option<u16> { ... }
}

Used to define a set of callbacks in your code. They would be called by the parser whenever new data is available. You should bear in mind that the data might get in your callbacks in a partial form.

Return Option as a result of each function call - either None for the "OK, go on" status, or Some(1) when you want to stop the parser after the function call is ended.

All callbacks provide a default no-op implementation (i.e. they just return None).

Provided Methods

fn on_url(&self, &String) -> Option<u16>

Called when the URL part of a request becomes available. E.g. for GET /forty-two HTTP/1.1 it will be called with "/forty_two" argument.

It's not called in the response mode.

fn on_status(&self, &String) -> Option<u16>

Called when a response status becomes available.

It's not called in the request mode.

fn on_header_field(&self, &String) -> Option<u16>

Called for each HTTP header key part.

fn on_header_value(&self, &String) -> Option<u16>

Called for each HTTP header value part.

fn on_body(&self, &String) -> Option<u16>

Called with body text as an argument when the new part becomes available.

fn on_headers_complete(&self) -> Option<u16>

Notified when all available headers have been processed.

fn on_message_begin(&self) -> Option<u16>

Notified when the parser receives first bytes to parse.

fn on_message_complete(&self) -> Option<u16>

Notified when the parser has finished its job.

fn on_chunk_header(&self) -> Option<u16>

fn on_chunk_complete(&self) -> Option<u16>